Refactorings.
The targets (both abstract and virtual) used to refer to project there belong
to using the name of Jamfile module corresponding to the project. That was a
bit messy, and whatevet it was necessary to get any project attribute, one
had to call 'project.attribute', which was messy too.

Now, projects are identified by instance of project-target class, and
attributes are obtained via 'get' method in that class.


git-svn-id: http://svn.boost.org/svn/boost/trunk/boost@23001 b8fc166d-592f-0410-95f2-cb63ce0dd405