Remove TODO note to self.
For the record: I removed the overload because it was causing an ambiguity that I was too lazy to work out at the time. It turned out this was because wchar_t was defined as unsigned short - using BOOST_NO_INTRINSIC_WCHAR_T was the right way to deal with this so the overload no longer needs to be disabled.


git-svn-id: http://svn.boost.org/svn/boost/trunk/boost@33114 b8fc166d-592f-0410-95f2-cb63ce0dd405