Quickbook: Parse lists with paragraphs.
Quite a big change. I moved the list logic into the grammar so that it's
easier to tell how to parse different blocks. Also reworked some of the
block vs. phrase stuff - it's a lot cleaner now which helped implement
this.  It generates terrible markup at the moment, but at least the
parser is in place.

git-svn-id: http://svn.boost.org/svn/boost/branches/quickbook-dev@75446 b8fc166d-592f-0410-95f2-cb63ce0dd405