Quickbook: Fix handling UTF-8 characters in the syntax highlighter.
So far, quickbook has just pretended that it's dealing with a single
byte encoding.  For the most part this works quite well due to the
design of UTF-8, but the syntax highlighter needs to pick out individual
characters correctly, so implement just enough to do that. Does odd
things with invalid UTF-8.

git-svn-id: http://svn.boost.org/svn/boost/branches/quickbook-dev@76790 b8fc166d-592f-0410-95f2-cb63ce0dd405