Appease clang for BOOST_NO_LOCAL_CLASS_TEMPLATE_PARAMETERS.
The use of a version check to determine the feature support is not recommended in clang. But we do the version check, since clang does not have a querying macro for this feature. This is an unfortunate workaround. We need to observe carefully the effect of the version check on vendor versions of clang (other than the LLVM project's clang).

git-svn-id: http://svn.boost.org/svn/boost/trunk@77649 b8fc166d-592f-0410-95f2-cb63ce0dd405